ml_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
Pillow==11.0.0
numpy==1.26.4
tensorflow==2.17.0
pydantic==2.9.2
